# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Учебный продакшн-подобный сервис учёта клиентов.\n\nКлючевая механика: обычный пользователь не редактирует карточку\nклиента напрямую — он ПРЕДЛАГАЕТ правку, которая применяется\nтолько после одобрения администратором.\n\nДемо-учётки (заводятся миграцией 00004):\nadmin@example.com / admin12345 — роль admin\nuser@example.com / user12345 — роль user",
        "title": "Clients API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Возвращает JWT для доступа к защищённым эндпоинтам.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Email и пароль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Не прошла валидация",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Неверный email или пароль",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает актуальные данные владельца токена из базы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Текущий пользователь",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Нет или невалиден токен",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь удалён",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создаёт пользователя с ролью user и сразу возвращает JWT.\nРоль admin через этот эндпоинт получить нельзя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Email и пароль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Не прошла валидация",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже занят",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vitikevich-landau_clients-api_internal_apperr.Code": {
            "type": "string",
            "enum": [
                "validation_error",
                "unauthorized",
                "forbidden",
                "not_found",
                "method_not_allowed",
                "conflict",
                "timeout",
                "internal_error"
            ],
            "x-enum-comments": {
                "CodeConflict": "нарушение уникальности / бизнес-правила",
                "CodeForbidden": "токен есть, но прав не хватает",
                "CodeInternal": "всё остальное, наша вина",
                "CodeMethodNotAllowed": "путь есть, метод не тот",
                "CodeNotFound": "объект не существует",
                "CodeTimeout": "не уложились в отведённое время",
                "CodeUnauthorized": "нет токена или он невалиден",
                "CodeValidation": "невалидные входные данные"
            },
            "x-enum-descriptions": [
                "невалидные входные данные",
                "нет токена или он невалиден",
                "токен есть, но прав не хватает",
                "объект не существует",
                "путь есть, метод не тот",
                "нарушение уникальности / бизнес-правила",
                "не уложились в отведённое время",
                "всё остальное, наша вина"
            ],
            "x-enum-varnames": [
                "CodeValidation",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeTimeout",
                "CodeInternal"
            ]
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "ExpiresIn — сколько СЕКУНД токен ещё живёт.\nОтносительное время, а не абсолютное: клиенту не нужно\nиметь правильно выставленные часы, чтобы понять срок.",
                    "type": "integer"
                },
                "token_type": {
                    "description": "TokenType всегда \"Bearer\" — подсказка клиенту, что токен\nнадо слать в заголовке Authorization: Bearer \u003ctoken\u003e.",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.UserResponse"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "description": "max=72 — это не каприз, а ограничение bcrypt: алгоритм работает\nмаксимум с 72 байтами. Более длинные пароли современные версии\nx/crypto отвергают ошибкой (старые молча обрезали, что было хуже:\nпароли \"…длинный…AAA\" и \"…длинный…BBB\" оказывались одинаковыми).",
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.Role": {
            "type": "string",
            "enum": [
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin"
            ]
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Role"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код: not_found, validation_error, ...\nКлиент принимает решения ПО КОДУ, а не по тексту:\nтекст можно переписать или перевести, код — контракт.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_apperr.Code"
                        }
                    ]
                },
                "details": {
                    "description": "Details — пояснения по конкретным полям при ошибке валидации.\nomitempty: в остальных случаях ключа в JSON просто не будет.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Message — человекочитаемое описание. Безопасное: без имён таблиц,\nSQL и стектрейсов.",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID — чтобы пользователь мог назвать его в поддержке,\nа мы по нему нашли весь запрос в логах.",
                    "type": "string"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorBody"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите: Bearer \u003cваш JWT\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}